+++ src/main/java/com/todaySee/home/config/HomeUserDetailsService.java
@@ -9,7 +9,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import com.todaySee.domain.UserVO;
-import com.todaySee.home.persistence.UserRepository;
+import com.todaySee.persistence.UserRepository;
 
 /*
  *  작성자 : 문혜진
