+++ src/main/webapp/WEB-INF/views/home/homeEmailFindList.jsp
@@ -16,7 +16,7 @@
 <meta name="viewport"
 	content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
-<title>ID 찾기</title>
+<title>Email 찾기</title>
 
 <!-- Favicon -->
 <link rel="shortcut icon" href="/images/home/favicon.png">
